+++ source/client.js
@@ -28,6 +28,7 @@ import { clearStore, createStore } from './store.js'
  * @property {boolean} isOnline - Currently online anywhere in Immers Space
  * @property {string} [locationName] - Name of current or last immer visited
  * @property {string} [locationURL] - URL of current or last immer visited
+ * @property {('friend-online'|'friend-offline'|'request-receved'|'request-sent'|'none')} status - descriptor of the current relationship to this user
  * @property {string} statusString - Text description of current status, "Offline" / "Online at..."
  * @property {string} __unsafeStatusHTML - Unsanitized HTML description of current status with link.
  * You must sanitize this string before inserting into the DOM to avoid XSS attacks.
@@ -309,7 +310,12 @@ export class ImmersClient extends window.EventTarget {
    */
   async friendsList () {
     const friendsCol = await this.activities.friends()
+    this.#store.friends = friendsCol.orderedItems
+      .map(ImmersClient.FriendStatusFromActivity)
     return friendsCol.orderedItems
+      // don't show ex-friends in list
+      .filter(activity => activity.type !== 'Reject')
+      // map it again to avoid shared, mutable objects
       .map(ImmersClient.FriendStatusFromActivity)
   }
 
@@ -338,11 +344,51 @@ export class ImmersClient extends window.EventTarget {
    * @param {string[]} [to] - Addressees. Accepts Immers handles (username[domain.name]) and ActivityPub IRIs
    * @returns {Promise<string>} Url of newly posted message
    */
-  async sendChatMessage (content, privacy, to = []) {
+  sendChatMessage (content, privacy, to = []) {
     return this.activities.note(DOMPurify.sanitize(content), to, privacy)
   }
 
   /**
+   * This method will either initiate a new friend request or,
+   * if a request has already been received from the target user,
+   * accept a pending friend request. To create a friend connection,
+   * both users will need to call this method with the other user's handle.
+   * @param  {string} handle - the target user's immers handle or profile id
+   */
+  async addFriend (handle) {
+    const userId = handle.startsWith('https://') ? handle : await this.resolveProfileIRI(handle)
+    const pendingRequest = this.#store.friends?.find(status => status.profile.id === userId && status.status === 'request-received')
+    if (pendingRequest) {
+      return this.activities.accept(pendingRequest._activity)
+    }
+    return this.activities.follow(userId)
+  }
+
+  /**
+   * Remove a relationship to another immerser,
+   * either by removing an existing friend,
+   * rejecting a pending incoming friend request,
+   * or canceling a pending outgoing friend request
+   * @param  {string} handle - the target user's immers handle or profile id
+   */
+  async removeFriend (handle) {
+    const userId = handle.startsWith('https://') ? handle : await this.resolveProfileIRI(handle)
+    const pendingRequest = this.#store.friends
+      ?.find(status => status.profile.id === userId && status.status === 'request-received')
+    if (pendingRequest) {
+      return this.activities.reject(pendingRequest._activity.id, userId)
+    }
+    const pendingOutgoingRequest = this.#store.friends
+      ?.find(status => status.profile.id === userId && status.status === 'request-sent')
+    if (pendingOutgoingRequest) {
+      return this.activities.undo(pendingOutgoingRequest._activity)
+    }
+    // technically reject needs the original follow activity ID, but
+    // immers server will do this lookup for us if we send reject of a friends list user id
+    return this.activities.reject(userId, userId)
+  }
+
+  /*
    * Upload a 3d model as an avatar and optionally share it
    * @param  {string} name - Name/description
    * @param  {Blob} glb - 3d model gltf binary file
@@ -572,24 +618,51 @@ export class ImmersClient extends window.EventTarget {
    * @returns {FriendStatus}
    */
   static FriendStatusFromActivity (activity) {
-    const isOnline = activity.type === 'Arrive'
     const locationName = activity.target?.name
     const locationURL = activity.target?.url
-    const statusString = isOnline
-      ? `Online at ${locationName} (${locationURL})`
-      : 'Offline'
-    const __unsafeStatusHTML = isOnline
-      ? `<span>Online at <a href="${locationURL}">${locationName}</a></span>`
-      : '<span>Offline</span>'
-    return {
-      profile: ImmersClient.ProfileFromActor(activity.actor),
+    let status = 'none'
+    let statusString = ''
+    let __unsafeStatusHTML = '<span></span>'
+    let actor = activity.actor
+    switch (activity.type.toLowerCase()) {
+      case 'arrive':
+        status = 'friend-online'
+        statusString = `Online at ${locationName} (${locationURL})`
+        __unsafeStatusHTML = `<span>Online at <a href="${locationURL}">${locationName}</a></span>`
+        break
+      case 'leave':
+      case 'accept':
+        status = 'friend-offline'
+        statusString = 'Offline'
+        __unsafeStatusHTML = `<span>${statusString}</span>`
+        break
+      case 'follow':
+        if (actor.id) {
+          status = 'request-received'
+          statusString = 'Sent you a friend request'
+          __unsafeStatusHTML = `<span>${statusString}</span>`
+        } else if (activity.object?.id) {
+          // for outgoing request, current user is the actor; we're interested in the object
+          actor = activity.object
+          status = 'request-sent'
+          statusString = 'You sent a friend request'
+          __unsafeStatusHTML = `<span>${statusString}</span>`
+        }
+        break
+    }
+    const isOnline = status === 'online'
+    const friendStatus = {
+      profile: ImmersClient.ProfileFromActor(actor),
       isOnline,
       locationName,
       locationURL,
+      status,
       statusString,
       __unsafeStatusHTML,
       statusHTML: DOMPurify.sanitize(__unsafeStatusHTML)
     }
+    Object.defineProperty(friendStatus, '_activity', { enumerable: false, value: activity })
+    return friendStatus
   }
 
   /**
@@ -690,11 +763,12 @@ export class ImmersClient extends window.EventTarget {
         headers: { Accept: Activities.JSONLDMime }
       }).then(res => res.json())
     } else {
+      const defaultPlace = { type: 'Place', audience: Activities.PublicAddress }
       const basePlace = this.localImmer
-        ? await window.fetch(`${this.localImmer}/o/immer`, {
+        ? await window.fetch(`${getURLPart(this.localImmer, 'origin')}/o/immer`, {
             headers: { Accept: Activities.JSONLDMime }
-          })
-        : { type: 'Place', audience: Activities.PublicAddress }
+          }).then(res => res.json()).catch(() => defaultPlace)
+        : defaultPlace
       this.place = Object.assign(
         basePlace,
         destinationDescription