+++ CHANGELOG.md
@@ -1,23 +1,36 @@
 ## Unreleased
 
+### Friend management
+
+New high-level apis to add & remove friends
+
+* New `ImmersClient.addFriend` - Send a friend request or accept one in order to create a relationship and start sharing location
+* New `ImmersClient.removeFriend` - Cancel a previous relationship or request
+* Updated `ImmersClient.friendsList` - exclude ex-friends
+* Updated `FriendStatus` structure - include `status` enum (friend-online, friend-offline, request-receved, request-sent, none) and user `profile`
+* Updated `ImmersClient` event `'immers-client-friends-update'` - also fires when an outgoing activity (e.g. friend request/accept) causes the statuses to update\*
+* Fixed destination object for local immer not fetched correctly
+
 ### Avatar utilities
 
-New low- and high-level apis for creating, saving, sharing, and using avatars.
-
-High-level apis
+New high-level apis for creating, saving, sharing, and using avatars.
 
 * New `ImmersClient.createAvatar` - uploads a 3D avatar and thumbnail and Creates a Model object for it\*
 * New `ImmersClient.addAvatar` - add an avatar to a user's portable avatar collection
 * new `ImmersClient.useAvatar` - change a users current avatar
 * Updated `ImmersClient.Profile` structure - include listing of user collections
 
-Low-level apis
+### Low-level apis
 
 * New `Activities.postMedia` - upload and share media files via ActivityPub Media Upload\*
 * New `Activities.model` - upload a 3D model with `postMedia`\*
+* New `Activities.undo` - revert a past activity, such as a Follow or Accept
 * Updated `Activities.add` - clarify inputs and allow creation of new collections
+* New `ImmersSocket` event `'immers-socket-outbox-update'` - when th current user has posted something new, either from current client or another\*
+* Fixed the auto leave-on-disconnect activity to have the same summary format as `ImmersClient.exit` activity
 
 \*  requires user's home Immers server version >= 2.2.0, otherwise error 'Missing/invalid upload media endpoint' will be thrown.
+
 ## v2.3.0
 
 * New method `ImmersClient.loginWithToken` - allow login without user interaction if credentials acquired through another means (for controlled accounts feature coming soon in immers server)